Semi-fixed flickering issue for the side nav menu
The problem was and still is only apparent on mobile, specifically iPhone (extended size variant). The flickering was worse but altering the placement of the black 'contact' sections on the E-board cards lessened the issue weirdly enough. Also enabled triggering of the footer and bottom nav bar whenever the side nav is opened in order to allow the viewer to view content on mobile unimpeded.